(BOLT-1358) Fix local_windows transport calling profile
Previous to this commit the local_windows transport would call the powershell
profile if the default value for the 'interpreter' variable was ever used. in
this scenario the default interpreter needs to include PS_ARGS defined in the
powershell transport.

This commit adds PS_ARGS to the default of interpreter. Also, in order to easilly access the default powershell arguments the PS_ARGS array has been converted to a module method.