(maint) Update resolver spec
The module resolver spec tests assert an error is raised when trying to resolve a module that is unresolvable. The tests use a "real" module for these tests. That module/version combination now has a circular dependency issue  which is producing a message that is different than what was previously expected in the tests. The important thing for this particular test is that a bolt error is raised (as even if the circular dependency issue is resolved by the module author, the particular version will still have an unknown module dependency). This commit updates the test to only assert that a bolt error is raised, so that if the circular dependency issue is fixed upstream the test will continue to work (by avoiding an overly specific error message match).